Issue #2583, Issue #2600 - move ConstraintFactory back into Constraint
